+++ src/gui/dialogs/gd_editor.cpp
@@ -202,7 +202,7 @@ gdEditor::gdEditor(SampleChannel *ch)
 
   if (ch->panRight < 1.0f) {
     char buf[8];
-    sprintf(buf, "%d L", abs((ch->panRight * 100.0f) - 100));
+    sprintf(buf, "%d L", (int) abs((ch->panRight * 100.0f) - 100));
     pan->value(ch->panRight);
     panNum->value(buf);
   }
@@ -212,7 +212,7 @@ gdEditor::gdEditor(SampleChannel *ch)
   }
   else {
     char buf[8];
-    sprintf(buf, "%d R", abs((ch->panLeft * 100.0f) - 100));
+    sprintf(buf, "%d R", (int) abs((ch->panLeft * 100.0f) - 100));
     pan->value(2.0f - ch->panLeft);
     panNum->value(buf);
   }
